Move DialURL into http/common
