+++ authority/voting/client/client.go
@@ -20,15 +20,11 @@ package client
 import (
 	"context"
 	"crypto/hmac"
-	"crypto/tls"
-	"crypto/x509"
 	"errors"
 	"fmt"
 	"net"
 	"net/url"
 
-	"github.com/quic-go/quic-go"
-	"github.com/quic-go/quic-go/http3"
 	"gopkg.in/op/go-logging.v1"
 
 	"github.com/katzenpost/katzenpost/authority/voting/server/config"
@@ -123,49 +119,6 @@ func newConnector(cfg *Config) *connector {
 	return p
 }
 
-func DialURL(u *url.URL, ctx context.Context, dialFn func(ctx context.Context, network, address string) (net.Conn, error)) (net.Conn, error) {
-	switch u.Scheme {
-	case "tcp":
-		// XXX: make sure to use the supplied dialer for proxy users
-		conn, err := dialFn(ctx, "tcp", u.Host)
-		if err != nil {
-			return nil, err
-		} else {
-			return conn, nil
-		}
-	case "http":
-		// http/3 quic connector
-		// XXX: will need to add the TLS certificate
-		// fingerprint to the authority configuration
-		// or obtain valid CA-signed certificates for
-		// the authorities.
-		pool, err := x509.SystemCertPool()
-		if err != nil {
-			panic(err)
-		}
-		tlsConf := &tls.Config{
-			RootCAs:            pool,
-			InsecureSkipVerify: true, // XXX
-			// ALPN is externally visible as part of the client/server hello,
-			// so pick a common protocol rather than something fingerprintable.
-			NextProtos: []string{http3.NextProtoH3},
-		}
-		qconn, err := quic.DialAddr(u.Host, tlsConf, nil)
-		if err == nil {
-			// open a quic stream
-			stream, err := qconn.OpenStreamSync(ctx)
-			if err == nil {
-				// wrap the stream and conn to implement net.Conn
-				return &common.QuicConn{Stream: stream, Conn: qconn}, nil
-			}
-			return nil, err
-		}
-		return nil, err
-	default:
-		return nil, errors.New("Unsupported Scheme")
-	}
-}
-
 func (p *connector) initSession(ctx context.Context, doneCh <-chan interface{}, linkKey wire.PrivateKey, signingKey sign.PublicKey, peer *config.Authority) (*connection, error) {
 	var conn net.Conn
 	var err error
@@ -186,7 +139,7 @@ func (p *connector) initSession(ctx context.Context, doneCh <-chan interface{}, 
 		if err != nil {
 			continue
 		}
-		conn, err = DialURL(u, ctx, dialFn)
+		conn, err = common.DialURL(u, ctx, dialFn)
 		if err == nil {
 			break
 		}
