+++ authority/voting/server/state.go
@@ -51,6 +51,7 @@ import (
 	"github.com/katzenpost/katzenpost/core/wire"
 	"github.com/katzenpost/katzenpost/core/wire/commands"
 	"github.com/katzenpost/katzenpost/core/worker"
+	"github.com/katzenpost/katzenpost/http/common"
 )
 
 const (
@@ -677,7 +678,7 @@ func (s *state) sendCommandToPeer(peer *config.Authority, cmd commands.Command) 
 			continue
 		}
 		defaultDialer := &net.Dialer{}
-		conn, err = client.DialURL(u, context.Background(), defaultDialer.DialContext)
+		conn, err = common.DialURL(u, context.Background(), defaultDialer.DialContext)
 		if err == nil {
 			break
 		}
